+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/components/input-document/input-document.context.ts
@@ -1,10 +1,7 @@
 import type { UmbDocumentPickerModalData, UmbDocumentPickerModalValue } from '../../modals/types.js';
-import {
-	UMB_DOCUMENT_PICKER_MODAL,
-	UMB_DOCUMENT_ITEM_REPOSITORY_ALIAS,
-	UMB_DOCUMENT_SEARCH_PROVIDER_ALIAS,
-} from '../../constants.js';
-import type { UmbDocumentItemModel } from '../../repository/index.js';
+import { UMB_DOCUMENT_PICKER_MODAL, UMB_DOCUMENT_SEARCH_PROVIDER_ALIAS } from '../../constants.js';
+import type { UmbDocumentItemModel } from '../../item/types.js';
+import { UMB_DOCUMENT_ITEM_REPOSITORY_ALIAS } from '../../item/constants.js';
 import type { UmbDocumentTreeItemModel } from '../../tree/types.js';
 import { UmbPickerInputContext } from '@umbraco-cms/backoffice/picker-input';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
