+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/constants.ts
@@ -3,6 +3,7 @@ export { UMB_DOCUMENT_ENTITY_TYPE, UMB_DOCUMENT_ROOT_ENTITY_TYPE } from './entit
 export * from './collection/constants.js';
 export * from './entity-actions/constants.js';
 export * from './entity-bulk-actions/constants.js';
+export * from './item/constants.js';
 export * from './modals/constants.js';
 export * from './paths.js';
 export * from './property-dataset-context/constants.js';
