+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/index.ts
@@ -5,16 +5,17 @@ export * from './components/index.js';
 export * from './constants.js';
 export * from './entity-actions/index.js';
 export * from './global-contexts/index.js';
+export * from './item/index.js';
 export * from './modals/index.js';
 export * from './paths.js';
 export * from './publishing/index.js';
 export * from './recycle-bin/index.js';
 export * from './reference/index.js';
 export * from './repository/index.js';
+export * from './tree/index.js';
 export * from './url/index.js';
 export * from './user-permissions/index.js';
 
-export * from './tree/index.js';
 export { UMB_CONTENT_MENU_ALIAS } from './menu/manifests.js';
 export { UMB_DOCUMENT_COLLECTION_ALIAS } from './collection/constants.js';
