+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/item/constants.ts
@@ -1,0 +1,1 @@
+export * from './repository/constants.js';
