+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/item/index.ts
@@ -1,0 +1,1 @@
+export { UmbDocumentItemRepository } from './repository/index.js';
