+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/item/repository/manifests.ts
