+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/item/types.ts
@@ -1,0 +1,1 @@
+export type { UmbDocumentItemModel } from './repository/types.js';
