+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/repository/constants.ts
@@ -1,3 +1,2 @@
 export * from './detail/constants.js';
-export * from './item/constants.js';
 export * from './validation/constants.js';
