+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/repository/index.ts
@@ -1,6 +1,4 @@
 export { UmbDocumentDetailRepository } from './detail/index.js';
-export { UmbDocumentItemRepository } from './item/index.js';
 export { UmbDocumentPreviewRepository } from './preview/index.js';
 
 export * from './constants.js';
-export type * from './types.js';
