+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/repository/manifests.ts
@@ -1,4 +1,3 @@
 import { manifests as detailManifests } from './detail/manifests.js';
-import { manifests as itemManifests } from './item/manifests.js';
 
-export const manifests: Array<UmbExtensionManifest> = [...detailManifests, ...itemManifests];
+export const manifests: Array<UmbExtensionManifest> = [...detailManifests];
