--- src/Umbraco.Web.UI.Client/src/packages/documents/documents/repository/types.ts
@@ -1,1 +1,0 @@
-export type { UmbDocumentItemModel } from './item/types.js';
