+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/types.ts
@@ -10,15 +10,15 @@ import type { UmbContentDetailModel, UmbElementValueModel } from '@umbraco-cms/b
 export { UmbDocumentVariantState };
 
 export type * from './audit-log/types.js';
-export type * from './recycle-bin/types.js';
 export type * from './collection/types.js';
+export type * from './entity.js';
+export type * from './item/types.js';
 export type * from './modals/types.js';
-export type * from './repository/types.js';
+export type * from './publishing/types.js';
+export type * from './recycle-bin/types.js';
 export type * from './tree/types.js';
 export type * from './user-permissions/types.js';
-export type * from './entity.js';
 export type * from './workspace/types.js';
-export type * from './publishing/types.js';
 
 export interface UmbDocumentDetailModel extends UmbContentDetailModel {
 	documentType: {
