+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/user-permissions/input-document-granular-user-permission/input-document-granular-user-permission.element.ts
@@ -1,5 +1,6 @@
 import type { UmbDocumentUserPermissionModel } from '../types.js';
-import { UmbDocumentItemRepository, type UmbDocumentItemModel } from '../../repository/index.js';
+import { UmbDocumentItemRepository } from '../../item/index.js';
+import type { UmbDocumentItemModel } from '../../item/types.js';
 import { UMB_DOCUMENT_PICKER_MODAL } from '../../constants.js';
 import { css, customElement, html, repeat, state } from '@umbraco-cms/backoffice/external/lit';
 import { UmbLitElement } from '@umbraco-cms/backoffice/lit-element';
